# API/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Serve",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://nicolas-camilloni.students-laplateforme.io/:8080",
            "description": "Mon API"
        }
    ],
    "paths": {
        "/api/users?login={login}": {
            "put": {
                "operationId": "App\\Controllers\\UserController::putUser",
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "description": "login à modifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "login",
                        "in": "path",
                        "description": "login à modifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Modification du login ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erreur ds la modification du login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}